Title: 공책과 포스트잇, Time: 46ms, Memory: 1MB, Status: Passed - CodeTree

+++ 250218/공책과 포스트잇/note-and-sticky-note.cpp
@@ -30,7 +30,7 @@ int main() {
 	long long ans = 0;
 
 	while (le <= re) {
-		int mid = (le + re) / 2;
+		long long mid = (le + re) / 2;
 
 		if (Calc(mid)) {
 			le = mid + 1;
